+++ make.rules
@@ -5,6 +5,7 @@ NOGZIPARCHIVES=1
 
 ARCHFILES=\
 	src/adlmidi.cpp src/adlmidi.h \
+    src/adlmidi_load.cpp src/adlmidi_midiplay.cpp src/adlmidi_opl3.cpp src/adlmidi_private.cpp \
 	src/dbopl.cpp src/dbopl.h \
 	src/adldata.cpp src/adldata.hh \
 	src/dumpbank/dumpbank.cc src/dumpmiles/dumpmiles.cc \
@@ -100,10 +101,22 @@ directories:
 	${MKDIR} obj
 	${MKDIR} bin
 
-libadlmidi.so: obj/adlmidi.o obj/nukedopl3.o obj/adldata.o
+libadlmidi.so: obj/adlmidi.o obj/adlmidi_load.o obj/adlmidi_midiplay.o obj/adlmidi_opl3.o obj/adlmidi_private.o obj/nukedopl3.o obj/adldata.o
 	$(CXXLINK)  $^  $(DEBUG) -shared -o bin/$@  $(LDLIBS)
 
-obj/adlmidi.o: src/adlmidi.cpp src/adlmidi.h src/dbopl.h src/adldata.hh
+obj/adlmidi.o: src/adlmidi.cpp src/adlmidi.h src/dbopl.h src/nukedopl3.h src/adldata.hh
+	$(CXX) $(CPPFLAGS) $<  $(DEBUG) -c -fPIC -o $@
+
+obj/adlmidi_load.o: src/adlmidi_load.cpp src/adlmidi_private.hpp src/adlmidi.h src/adldata.hh
+	$(CXX) $(CPPFLAGS) $<  $(DEBUG) -c -fPIC -o $@
+
+obj/adlmidi_midiplay.o: src/adlmidi_midiplay.cpp src/adlmidi_private.hpp src/adlmidi.h src/adldata.hh
+	$(CXX) $(CPPFLAGS) $<  $(DEBUG) -c -fPIC -o $@
+
+obj/adlmidi_opl3.o: src/adlmidi_opl3.cpp src/adlmidi_private.hpp src/adlmidi.h src/adldata.hh
+	$(CXX) $(CPPFLAGS) $<  $(DEBUG) -c -fPIC -o $@
+
+obj/adlmidi_private.o: src/adlmidi_private.cpp src/adlmidi_private.hpp src/adlmidi.h src/adldata.hh
 	$(CXX) $(CPPFLAGS) $<  $(DEBUG) -c -fPIC -o $@
 
 obj/dbopl.o: src/dbopl.cpp src/dbopl.h
@@ -113,25 +126,25 @@ obj/nukedopl3.o: src/nukedopl3.c src/nukedopl3.h
 	$(CC) $(CFLAGS) $<  $(DEBUG)  -c -fPIC -o $@
 
 obj/adldata.o: src/adldata.cpp src/adldata.hh
-	$(CXX) $(CPPFLAGS) $<  $(DEBUG)  -c -fPIC -o $@ 
+	$(CXX) $(CPPFLAGS) $<  $(DEBUG)  -c -fPIC -o $@
 
 gen_adldata: obj/gen_adldata.o obj/dbopl.o
 	$(CXXLINK)  $^  $(DEBUG)  -o bin/$@  $(LDLIBS)
 
 obj/gen_adldata.o: src/gen_adldata/gen_adldata.cc src/dbopl.h
-	$(CXX) $(CPPFLAGS) $<  $(DEBUG)  -c -o $@ 
+	$(CXX) $(CPPFLAGS) $<  $(DEBUG)  -c -o $@
 
 dumpmiles: obj/dumpmiles.o
 	$(CXXLINK)  $^  $(DEBUG)  -o bin/$@  $(LDLIBS)
 
 obj/dumpmiles.o: src/dumpmiles/dumpmiles.cc
-	$(CXX) $(CPPFLAGS) $<  $(DEBUG)  -c -o $@ 
+	$(CXX) $(CPPFLAGS) $<  $(DEBUG)  -c -o $@
 
 dumpbank: obj/dumpbank.o
 	$(CXXLINK)  $^  $(DEBUG)  -o bin/$@  $(LDLIBS)
 
 obj/dumpbank.o: src/dumpbank/dumpbank.cc
-	$(CXX) $(CPPFLAGS) $<  $(DEBUG)  -c -o $@ 
+	$(CXX) $(CPPFLAGS) $<  $(DEBUG)  -c -o $@
 
 .PHONY: clean
